add thumbs; fix thumb name; grid vs list; custom button; scroll space below recylcer
