+++ app/build.gradle
@@ -76,4 +76,7 @@ dependencies {
     implementation 'com.github.javafaker:javafaker:1.0.2'
 
     implementation 'com.google.android.play:core:1.10.0'
+
+    implementation 'androidx.cardview:cardview:latest_version'
+
 }
