+++ app/src/main/java/com/michaelaskew/avocadotimer/activities/AvocadoListActivity.java
@@ -1,12 +1,15 @@
 package com.michaelaskew.avocadotimer.activities;
 
 import android.os.Bundle;
+import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.michaelaskew.avocadotimer.R;
+import com.michaelaskew.avocadotimer.database.DatabaseHelper;
 import com.michaelaskew.avocadotimer.models.Avocado;
 
 import java.util.ArrayList;
@@ -27,7 +30,42 @@ public class AvocadoListActivity extends AppCompatActivity {
 
         loadAvocados(); // Move the code for loading avocado data here
 
-        // ... Rest of your RecyclerView-related code
+        ItemTouchHelper.SimpleCallback simpleItemTouchCallback = new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT) {
+
+            @Override
+            public boolean onMove(RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, RecyclerView.ViewHolder target) {
+                return false;
+            }
+
+            @Override
+            public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction) {
+                int position = viewHolder.getAdapterPosition();
+
+                if (position != RecyclerView.NO_POSITION && position < avocadoList.size()) {
+
+                    Avocado avocadoToDelete = avocadoList.get(position);
+
+                    DatabaseHelper dbHelper = new DatabaseHelper(AvocadoListActivity.this);
+                    if (dbHelper.deleteAvocado(avocadoToDelete.getId())) {
+                        avocadoList.remove(position);
+
+                        // Check if the list is empty after removing the item
+                        if (avocadoList.isEmpty()) {
+                            rvAvocadoList.getAdapter().notifyDataSetChanged();
+                        } else {
+                            rvAvocadoList.getAdapter().notifyItemRemoved(position);
+                        }
+
+                        Toast.makeText(AvocadoListActivity.this, "Avocado deleted!", Toast.LENGTH_SHORT).show();
+                    } else {
+                        Toast.makeText(AvocadoListActivity.this, "Error deleting Avocado.", Toast.LENGTH_SHORT).show();
+                    }
+                }
+            }
+        };
+
+        ItemTouchHelper itemTouchHelper = new ItemTouchHelper(simpleItemTouchCallback);
+        itemTouchHelper.attachToRecyclerView(rvAvocadoList);
     }
 
     private void loadAvocados() {
