+++ app/src/main/java/com/michaelaskew/avocadotimer/adapters/AvocadoAdapter.java
@@ -2,11 +2,14 @@ package com.michaelaskew.avocadotimer.adapters;
 
 import android.content.Context;
 import android.content.Intent;
+import android.net.Uri;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
@@ -33,7 +36,7 @@ public class AvocadoAdapter extends RecyclerView.Adapter<AvocadoAdapter.AvocadoV
     @NonNull
     @Override
     public AvocadoViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(context).inflate(R.layout.avocado_list_item, parent, false);
+        View view = LayoutInflater.from(context).inflate(R.layout.avocado_grid_item, parent, false);
         return new AvocadoViewHolder(view);
     }
 
@@ -43,17 +46,27 @@ public class AvocadoAdapter extends RecyclerView.Adapter<AvocadoAdapter.AvocadoV
             return;  // Safety check
         }
         Avocado avocado = avocadoList.get(position);
+        Uri thumb = Uri.parse(avocado.getImagePath());
 
         boolean nameIsSet = !avocado.getName().isEmpty();
         String creationTime = avocado.getCreationTime();
         int squishiness = avocado.getSquishiness();
+        holder.tvThumbnail.setImageURI(thumb);
         holder.tvName.setText(nameIsSet ? avocado.getName() : "[No Name]");
-
         holder.tvCreatedAt.setText(TimeUtils.getRelativeTimeText(creationTime));
 
         double fractionElapsed = (double) TimeUtils.getTimeRemaining(creationTime, 360, squishiness)[1]; // Assuming this method returns the correct value
         holder.circleChartView.setFractionElapsed(fractionElapsed);  // Set the fraction elapsed to the CircleChartView of the current item
 
+        // Post a runnable to get width after the layout is complete
+        holder.tvThumbnail.post(new Runnable() {
+            @Override
+            public void run() {
+                int width = holder.tvThumbnail.getWidth();
+                FrameLayout.LayoutParams layoutParams = new FrameLayout.LayoutParams(width, width);
+                holder.tvThumbnail.setLayoutParams(layoutParams);
+            }
+        });
         // ... (set other attributes of avocado)
         // ... (set other attributes of avocado)
     }
@@ -64,7 +77,7 @@ public class AvocadoAdapter extends RecyclerView.Adapter<AvocadoAdapter.AvocadoV
     }
 
     public class AvocadoViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
-
+        ImageView tvThumbnail;
         TextView tvName;
         TextView tvCreatedAt;
         CircleChartView circleChartView;
@@ -73,6 +86,7 @@ public class AvocadoAdapter extends RecyclerView.Adapter<AvocadoAdapter.AvocadoV
         public AvocadoViewHolder(@NonNull View itemView) {
             super(itemView);
             itemView.setOnClickListener(this);
+            tvThumbnail = itemView.findViewById(R.id.avocado_thumbnail);
             tvName = itemView.findViewById(R.id.avocado_name);
             tvCreatedAt = itemView.findViewById(R.id.avocado_creation_date);
             circleChartView = itemView.findViewById(R.id.circleChartView); // Initialize the CircleChartView
