+++ app/src/main/res/values/styles.xml
@@ -8,6 +8,15 @@
         <item name="android:statusBarColor">@color/colorPrimary</item>
     </style>
 
+    <style name="CustomButtonStyle">
+        <!-- Text properties -->
+        <item name="android:fontFamily">@font/sora_semibold</item>
+        <item name="android:textStyle">normal</item>
+        <item name="android:textSize">16sp</item>
+        <item name="android:letterSpacing">0.0</item>
+        <item name="android:textColor">#ffffff</item>
+        <item name="android:textAllCaps">false</item>
+    </style>
     <!-- Define colors for your app -->
     <color name="colorPrimary">#315624</color> <!-- Avocado green -->
     <color name="colorPrimaryDark">#213d17</color> <!-- Darker avocado green -->
